refactor: move archive workspace outside
